pyproject.toml
[project]
name = "core-nest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.2",
    "duo-orm-preview",
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "invoke==2.2.0",
    "litellm==1.83.0",
    "loguru>=0.7.3",
    "psycopg[binary,pool]==3.1.18",
    "pydantic==2.12.5",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "sqlalchemy==2.0.39",
    "tenacity==9.0.0",
    "uvicorn==0.34.0",
]

[tool.uv.sources]
duo-orm-preview = { git = "https://github.com/SiddhanthNB/duo-orm-preview.git", rev = "v0.1.1" }

[tool.duo-orm]
db_dir = "app"

[tool.duo-orm.migration]
version_table = "corenest__alembic_version"

[dependency-groups]
dev = [
    "fakeredis>=2.34.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]
